backend/requirements.txt
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
certifi==2022.12.7
click==8.1.3
DAWG-Python==0.7.2
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-cache2==0.2.1
fastapi-limiter==0.1.5
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.10
packaging==23.1
pendulum==2.1.2
pluggy==1.0.0
pydantic==1.10.7
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytzdata==2020.1
PyYAML==6.0
redis==4.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
